Installs Knative the thorough way,
with focus on Eventing.
Knative Buld is excluded because it is being phased out in favor of Tekton.

The Serving domain template is overridden with one that avoids subdomains,
to avoid headache for those of us who rely on wildcard certs.

Using unreleased versions because Knative Serving recently bumped to v1beta1 api
and Eventing's Kafka contrib got a major overhaul since the previous release.