The issue with go-containerregistry based components seemingly idling
for ~5 minutes before pull or digest check is quite common.
For example many, but not all, Skaffold builds get these delay after
the build script has exited (with zero CPU usage and zero subprocessed).
The workaround is to use :80 with our registry hosts in build steps,
but build steps only. Means we need to split skaffold run into build + deploy.
We rarely have the issue during development so maybe that's not a problem.