Downgrade Triggers until we have time to deal with long pod names
resulting from our use of $(uid) with tektoncd/triggers#926.

Long pod names work, but usability suffers. In particular with
tasks in a pipeline we no longer see the task name in kubectl watch.

For Dashboard it's also impractical because the pod name is the
headline of the taskrun page.

A solution could be based on generateName, like Triggers' PR 939.