buildkit >0.12.2 won't push to http=true registry with insecure=true
Failes with error: dial tcp [ip]:443: connect: connection refused

This is due to the new if-else in moby/buildkit#4299.
Thus we can no longer blindly add that output flag. We don't need it
anymore because the switch to http works in all our contexts,
including over kubefwd after we stopped exposing the https port.